+++ src/store/authors/authors.actions.js
@@ -1,10 +1,10 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { apiService } from '../../sesrvice';
+import { authorsService } from '../../services/authors.service';
 
 const fetchAllAuthors = createAsyncThunk(
 	'authors/fetchAllAuthors',
 	async () => {
-		const response = await apiService.getAllAuthors('authors/all');
+		const response = await authorsService.getAllAuthors('authors/all');
 		const result = await response.json();
 		return result.result;
 	}
