+++ src/store/user/actions.user.js
@@ -1,11 +1,11 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { apiService } from '../../sesrvice';
+import { usersService } from '../../services/users.sesrvice';
 
 const fetchLogin = createAsyncThunk(
 	'courses/fetchLogin',
 	async (body, { rejectWithValue }) => {
 		try {
-			const response = await apiService.login('login', body);
+			const response = await usersService.login('login', body);
 			const result = await response.json();
 			if (!response.ok) {
 				throw new Error(`${result.errors[0]}`);
@@ -21,7 +21,7 @@ const fetchRegistration = createAsyncThunk(
 	'courses/fetchRegistration',
 	async (body, { rejectWithValue }) => {
 		try {
-			const response = await apiService.registration('register', body);
+			const response = await usersService.registration('register', body);
 			const result = await response.json();
 			if (!response.ok) {
 				throw new Error(`${result.errors[0]}`);
@@ -34,7 +34,7 @@ const fetchRegistration = createAsyncThunk(
 );
 
 const fetchUsersMe = createAsyncThunk('courses/fetchUsersMe', async () => {
-	const response = await apiService.usersMe('users/me');
+	const response = await usersService.usersMe('users/me');
 	const result = await response.json();
 	return result;
 });
